Add support for customizing calls to HTCondor binaries to the HTCondor manager
Define the optional parameters `prefix`, `condor_rm_cmd` and `condor_submit_cmd` for the HTCondor job manager. The value of `prefix` is prepended to all calls to HTCondor binaries (i.e. `condor_rm` and `condor_submit`). The values of `condor_rm_cmd` and `condor_submit_cmd` control what binaries are called to remove jobs from the queue and submit jobs to the queue.